requirements.txt
# Python dispatch engine — no external dependencies required

# Web backend (backend/app.py)
flask>=3.0
flask-cors>=4.0
flask-socketio>=5.3
PyJWT>=2.8
bcrypt>=4.1
python-dotenv>=1.0
